feat: 🎸 upgrade to .net 6.0
