+++ source/RandomAvatar.sln
@@ -1,7 +1,7 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.29424.173
+# Visual Studio Version 17
+VisualStudioVersion = 17.3.32804.467
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{69F86E73-7AEF-479F-ABCF-C3B1FA527281}"
 	ProjectSection(SolutionItems) = preProject
@@ -9,7 +9,7 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 		..\README.md = ..\README.md
 	EndProjectSection
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "random-avatar", "random-avatar\random-avatar.csproj", "{3A245065-281F-47FB-8709-A030ABAA07EE}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "RandomAvatar", "RandomAvatar\RandomAvatar.csproj", "{F7B972F5-8CEE-4D2B-96E3-C5091A7524D3}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -17,10 +17,10 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{3A245065-281F-47FB-8709-A030ABAA07EE}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{3A245065-281F-47FB-8709-A030ABAA07EE}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{3A245065-281F-47FB-8709-A030ABAA07EE}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{3A245065-281F-47FB-8709-A030ABAA07EE}.Release|Any CPU.Build.0 = Release|Any CPU
+		{F7B972F5-8CEE-4D2B-96E3-C5091A7524D3}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{F7B972F5-8CEE-4D2B-96E3-C5091A7524D3}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{F7B972F5-8CEE-4D2B-96E3-C5091A7524D3}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{F7B972F5-8CEE-4D2B-96E3-C5091A7524D3}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
