+++ source/RandomAvatar/Properties/launchSettings.json
