+++ source/RandomAvatar/RandomAvatar.csproj
@@ -1,0 +1,17 @@
+﻿<Project Sdk="Microsoft.NET.Sdk.Web">
+
+  <PropertyGroup>
+    <TargetFramework>net6.0</TargetFramework>
+	  <ImplicitUsings>enable</ImplicitUsings>
+	  <Nullable>annotations</Nullable>
+	  <RootNamespace>RandomAvatar</RootNamespace>
+  </PropertyGroup>
+
+
+  <ItemGroup>
+ 
+	  <PackageReference Include="System.Drawing.Common" Version="6.0.0" />
+ 
+  </ItemGroup>
+
+</Project>
