+++ source/RandomAvatar/RandomAvatarException.cs
@@ -1,0 +1,12 @@
+﻿namespace RandomAvatar;
+
+public class RandomAvatarException : Exception
+{
+    public RandomAvatarException(Exception ex) : base("", ex)
+    {
+    }
+
+    public RandomAvatarException(string message) : base(message)
+    {
+    }
+}
