--- source/random-avatar/RandomAvatarException.cs
@@ -1,15 +1,0 @@
-﻿using System;
-
-namespace RandomAvatar
-{
-    public class RandomAvatarException : Exception
-    {
-        public RandomAvatarException(Exception ex) : base("", ex)
-        {
-        }
-
-        public RandomAvatarException(string message) : base(message)
-        {
-        }
-    }
-}
