--- source/random-avatar/random-avatar.csproj
@@ -1,16 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk.Web">
-
-  <PropertyGroup>
-    <TargetFramework>netcoreapp2.2</TargetFramework>
-  </PropertyGroup>
-
-
-  <ItemGroup>
- 
-    <PackageReference Include="Microsoft.AspNetCore.App" />
- 
-    <PackageReference Include="System.Drawing.Common" Version="4.6.0" />
- 
-  </ItemGroup>
-
-</Project>
